Fix incorrect keyset when there are no keypairs (#3)
If there are no keypairs and Keypair.keyset is called twice sequentially,
it will create keypairs on both calls. Additionally, in the first call
it will not return the future keypair, resulting in only one keypair
being returned. Only on the second call of Keypair.keyset will a future
keypair also be returned.